Parking  IOS Application
Over the decades our country has been developed drastically, now we are in this state that we have a lot of well contacted roads, commercial building and increasing number of automobiles. While parking these automobiles in parking space we are using the manual procedure of parking. Which most of the cases is unplanned and lack of discipline due to this, people can park their cars anywhere they want to, which creates a mess as people do not follow the particular cue most of the time. As a result of this, a hugecongestion  takes place in the parking area. While parking in and retrieving car due mismanagement cars can get dent by bumping with each other as there is lack of sufficient space. This leads to arguments, fights among people which sometimes makes huge traffic jam. This is also an economical loss as we need to repair our damaged car and also cars consume extra fuel while parking in or out. Traffic jam is an issue here as it kills our precious time. Due to this chaos in parking our valuable time gets wasted. It harms the students, office going staffs and emergency patients to a great extent
                                                         Figure1.1.1 :Existing car parking
It also causes economical loss to commercial places like shopping malls, amusement parks as people are more likely not to visit these places due to this parking hazard. As we are advancing with time, the manual car parking system in commercial spaces is creating hurdle which is causing wastage of time and some economic losses as well. Therefore we need a solution which can overcome these problems. Here we are introducing Automated Multilevel Car Parking Systems as a solution of these problems as well ascar parking system in commercial spaces is creating hurdle which is causing wastage of time and some economical losses as well. Therefore we need a solution which can overcome these problems. Here we are introducing AutomatedMultilevel Car Parking Systems as a solution of these problems as well as a replacement to the manual car parking systems at commercial spaces. This system not only saves time and money, it can also earn money by charging for parking spaces.